Delete TODO cache comment
I thought of creating cache while I was writing the parser, so it
possibly would take less time to fetch the message from configuration.
But then it turned out that if you parse a message before the
LocalizedMessage#set(String, String) method you'll have to deal with
reflection and a lot more code in the #set method, so I decided to
let it parse just before sending the message and cache BaseComponent[]
internally until rawMessage is changed.